pyproject.toml
[tool.poetry]
name = "refstudio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9, <3.12"  # pyinstaller requires Python <3.12
grobid-tei-xml = "^0.1.3"
grobid-client-python = "^0.0.5"
openai = "^0.27.7"
python-dotenv = "^1.0.0"
rank-bm25 = "^0.2.2"

[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
ruff = "^0.0.270"
pyinstaller = "^5.11.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120  # same length we have prettier/JS set to

[tool.coverage.run]
source = ["python"]
omit = ["*/tests/*"]